Some code refactorations
Moving struct private fields declaration from .c to a .h specific for private declarations. This change is necessary to make it easier and cleaner implementing unit tests. Also making some static functions inline and adding/improving some commnets
